+++ lib/src/components/common_counter/index.dart
@@ -1,0 +1,1 @@
+export 'common_counter.dart';